+++ tests/unittests/command_parse/test_geo.py
@@ -21,3 +21,14 @@ def test_geoadd(judge_command):
             "member": '"Catania"',
         },
     )
+
+
+def test_geosearch(judge_command):
+    judge_command(
+        "GEOSEARCH Sicily FROMLONLAT 15 37 BYBOX 400 400 km ASC WITHCOORD WITHDIST",
+        {
+            "command": "GEOSEARCH",
+            "key": "Sicily",
+            "any": "FROMLONLAT 15 37 BYBOX 400 400 km ASC WITHCOORD WITHDIST",
+        },
+    )